Added setpoint speed bootleg
It works to set the speed to a certain speed within limits, no
authorities being checked yet.  Also can simulate a failure, basically
just stops the train.  Still a bootleg, because I'm not sure how the
actual equation is supposed to work.  Either way, It sets the speed to
the right value pretty much.